+++ library/ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Features:
 - Added `Float32`, `Float64` types support
+- Added `Int64` settings type support
 
 # 1.0.0 -- 14-12-2025